Actually implement the changes lol
    Implements storing menus as arrays then loading those instead of doing unneeded IOs.
    Added some fail safes for invalided menus.
    Added more debugs.
